Add client operations interface, move and rename client operations implementation class
